[Misc] Separate out `utils.counter`
