+++ vllm/utils/__init__.py
@@ -3,7 +3,6 @@
 
 import enum
 import inspect
-import threading
 import uuid
 import warnings
 from functools import wraps
@@ -76,44 +75,6 @@ class Device(enum.Enum):
 class LayerBlockType(enum.Enum):
     attention = "attention"
     mamba = "mamba"
-
-
-class Counter:
-    def __init__(self, start: int = 0) -> None:
-        self.counter = start
-
-    def __next__(self) -> int:
-        i = self.counter
-        self.counter += 1
-        return i
-
-    def reset(self) -> None:
-        self.counter = 0
-
-
-class AtomicCounter:
-    """An atomic, thread-safe counter"""
-
-    def __init__(self, initial=0):
-        """Initialize a new atomic counter to given initial value"""
-        self._value = initial
-        self._lock = threading.Lock()
-
-    def inc(self, num=1):
-        """Atomically increment the counter by num and return the new value"""
-        with self._lock:
-            self._value += num
-            return self._value
-
-    def dec(self, num=1):
-        """Atomically decrement the counter by num and return the new value"""
-        with self._lock:
-            self._value -= num
-            return self._value
-
-    @property
-    def value(self):
-        return self._value
 
 
 def random_uuid() -> str:
